@@ -36,10 +36,16 @@ classifiers = [
 [tool.poetry.scripts]
 ipq = "ipq.__main__:main"
 
+[tool.poetry.extras]
+speedups = ["aiodns", "cchardet"]
+default = []
+
 [tool.poetry.dependencies]
-python = ">=3.6.2,<3.12"
+python = ">=3.7,<3.12"
 aiohttp = "^3.8.1"
 click = "^8.0.3"
+aiodns = { version = "^3.0.0", optional = true }
+cchardet = { version = "^2.1.7", optional = true }
 
 [tool.poetry.dev-dependencies]
 len8 = "^0.7.0"
